Window disassociation

+++ src/model/Mesh.cpp
@@ -4,7 +4,7 @@ Mesh::Mesh(const float tab_vect[][2], int size_tab_vect, const int tab_tri[][3],
 {
     for (int i = 0; i < size_tab_vect; i++)
     {
-        Vector2D vertex = Vector2D{static_cast<int>(tab_vect[i][0]), static_cast<int>(tab_vect[i][1])};
+        Vector2D vertex = Vector2D{tab_vect[i][0], tab_vect[i][1]};
         vertices.push_back(vertex);
     }
